+++ app/Http/Controllers/Public/CocktailController.php
@@ -6,6 +6,8 @@ namespace Kami\Cocktail\Http\Controllers\Public;
 
 use Illuminate\Http\Request;
 use Kami\Cocktail\Models\Bar;
+use OpenApi\Attributes as OAT;
+use Kami\Cocktail\OpenAPI as BAO;
 use Kami\Cocktail\Models\Cocktail;
 use Illuminate\Support\Facades\Cache;
 use Kami\Cocktail\Http\Controllers\Controller;
@@ -16,6 +18,22 @@ use Kami\Cocktail\Http\Resources\Public\CocktailResource;
 
 class CocktailController extends Controller
 {
+    #[OAT\Get(path: '/public/{barId}/cocktails', tags: ['Public'], operationId: 'listPublicBarCocktails', description: 'List and filter bar cocktails. To access this endpoint the bar must be marked as public.', summary: 'List cocktails', parameters: [
+        new OAT\Parameter(name: 'barId', in: 'path', required: true, description: 'Database id of bar', schema: new OAT\Schema(type: 'number')),
+        new BAO\Parameters\PageParameter(),
+        new OAT\Parameter(name: 'filter', in: 'query', description: 'Filter by attributes. You can specify multiple matching filter values by passing a comma separated list of values.', explode: true, style: 'deepObject', schema: new OAT\Schema(type: 'object', properties: [
+            new OAT\Property(property: 'name', type: 'string', description: 'Filter by cocktail names(s) (fuzzy search)'),
+            new OAT\Property(property: 'ingredient_name', type: 'string', description: 'Filter by cocktail ingredient names(s) (fuzzy search)'),
+            new OAT\Property(property: 'bar_shelf', type: 'boolean', description: 'Show only cocktails on the bar shelf'),
+            new OAT\Property(property: 'abv_min', type: 'number', description: 'Filter by greater than or equal ABV'),
+            new OAT\Property(property: 'abv_max', type: 'number', description: 'Filter by less than or equal ABV'),
+        ])),
+        new OAT\Parameter(name: 'sort', in: 'query', description: 'Sort by attributes. Available attributes: `name`, `created_at`, `average_rating`, `user_rating`, `abv`, `total_ingredients`, `missing_ingredients`, `missing_bar_ingredients`, `favorited_at`, `random`.', schema: new OAT\Schema(type: 'string')),
+    ], security: [])]
+    #[BAO\SuccessfulResponse(content: [
+        new BAO\PaginateData(CocktailResource::class),
+    ])]
+    #[BAO\NotFoundResponse]
     public function index(Request $request, int $barId): JsonResource
     {
         $bar = Bar::findOrFail($barId);
@@ -45,17 +63,30 @@ class CocktailController extends Controller
         return CocktailResource::collection($cocktails->withQueryString());
     }
 
+    #[OAT\Get(path: '/public/{barId}/cocktails/{slugOrPublicId}', tags: ['Public'], operationId: 'showPublicBarCocktail', description: 'Show public information about cocktail. If valid public ID is provided it will used, if not it will use cocktail slug.', summary: 'Show cocktail', parameters: [
+        new OAT\Parameter(name: 'barId', in: 'path', required: true, description: 'Database id of bar', schema: new OAT\Schema(type: 'number')),
+        new OAT\Parameter(name: 'slugOrPublicId', in: 'path', required: true, description: 'Cocktail slug or public id (ULID)', schema: new OAT\Schema(type: 'string')),
+    ], security: [])]
+    #[BAO\SuccessfulResponse(content: [
+        new BAO\WrapObjectWithData(CocktailResource::class),
+    ])]
+    #[BAO\NotFoundResponse]
     public function show(int $barId, string $slugOrPublicId): CocktailResource
     {
+        $cocktail = Cocktail::where('bar_id', $barId)
+            ->where('public_id', $slugOrPublicId)
+            ->orWhere('slug', $slugOrPublicId)
+            ->with('ingredients.ingredient', 'ingredients.substitutes.ingredient', 'images', 'tags', 'utensils')
+            ->firstOrFail();
+
+        if ($cocktail->public_id === $slugOrPublicId) {
+            return new CocktailResource($cocktail);
+        }
+
         $bar = Bar::findOrFail($barId);
         if (!$bar->is_public) {
             abort(404);
         }
-
-        $cocktail = Cocktail::where('public_id', $slugOrPublicId)
-            ->orWhere('slug', $slugOrPublicId)
-            ->firstOrFail()
-            ->load('ingredients.ingredient', 'ingredients.substitutes.ingredient', 'images', 'tags', 'utensils');
 
         return new CocktailResource($cocktail);
     }